+++ adkg/Cargo.toml
@@ -8,4 +8,12 @@ edition = "2021"
 avss = { path = "../avss" }
 util = { path = "../util" }
 rand = "0.8.5"
-sha256 = "1.4.0"
+colored = "2.0.0"
+termcolor = "1.1.2"
+
+[dev-dependencies]
+criterion = "0.4.0"
+
+[[bench]]
+name = "bench"
+harness = false
